+++ src/lua/lua_config.c
@@ -223,10 +223,11 @@ static int l_hin_create_log (lua_State *L) {
 }
 
 static int l_hin_redirect_log (lua_State *L) {
-  const char * path = lua_tostring (L, 1);
-  FILE * fp = NULL;
-
-  if (path) {
+  int type;
+  type = lua_type (L, 1);
+  if (type == LUA_TSTRING) {
+    const char * path = lua_tostring (L, 1);
+    FILE * fp = NULL;
     fp = freopen (path, "w", stdout);
     if (fp == NULL) { printf ("can't open '%s'\n", path); return 0; }
     fp = freopen (path, "w", stderr);
@@ -235,19 +236,20 @@ static int l_hin_redirect_log (lua_State *L) {
     dup2 (fileno(stderr), fileno(stdout));
     setvbuf (stdout, NULL, _IOLBF, 2024);
     setvbuf (stderr, NULL, _IOLBF, 2024);
-  } else if (!lua_isnil (L, 1)) {
+  } else if (type != LUA_TNONE && type != LUA_TNIL) {
     printf ("redirect log path needs to be a string\n");
   }
 
-  const char * mask = lua_tostring (L, 2);
-  if (mask) {
+  type = lua_type (L, 2);
+  if (type == LUA_TSTRING) {
+    const char * mask = lua_tostring (L, 2);
     char * ptr = NULL;
     uint32_t nr = strtol (mask, &ptr, 16);
     if (ptr <= mask) {
       printf ("can't parse debug mask\n");
     }
     master.debug = nr;
-  } else if (!lua_isnil (L, 2)) {
+  } else if (type != LUA_TNONE && type != LUA_TNIL) {
     printf ("redirect log debug mask needs to be a string\n");
   }